inventory-management/Backend/requirements.txt
aiosmtplib==2.0.2
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.6.2.post1
asyncpg==0.29.0
blinker==1.8.2
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.2
fastapi-mail==1.4.1
h11==0.14.0
idna==3.10
iso8601==2.1.0
Jinja2==3.1.4
MarkupSafe==3.0.2
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
pypika-tortoise==0.2.1
python-dotenv==1.0.1
pytz==2024.2
sniffio==1.3.1
starlette==0.40.0
tortoise-orm==0.21.7
typing_extensions==4.12.2
uvicorn==0.32.0
